requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.1
pycodestyle==2.8.0
pytz==2022.5
sqlparse==0.4.3
toml==0.10.2
whitenoise==5.3.0
